Enable screenshotting and editting of needles.
This commit adds support for Libvirt connections
that enable to take a screenshot from the virtual
machine running in the background, saving it in
the working directory (where the edited script)
is saved.

It also utilises the Needly application to allow
the users to edit needles directly without
having to run the Needly application later. However,
it they wish to do so, they still can can, because
the application runs on demand only.